Implement patch_bag as a first step.
This is rather noisy at the moment, but wanted to prove
the concept before determining the most clean and clear way
of doing things.

HTTP428 is not used yet, but will be used with tiddlers. See:
http://tools.ietf.org/html/draft-nottingham-http-new-status-04#section-3

The idea is that you shouldn't PATCH a tiddler unless you know
the Etag (implying that you know the content). Might need to revist
that idea depending on how people want to use it.

The same thing is quite a bit harder to do for bags and recipes as
Etag handling is not as robust.

When PATCHing a bag we can either reset the desc, or reset any
of the policy constraints. Since there are only two top level
keys on a bag, there's little point trying to make the code
for translating the HTTP request into a modified bag fully declarative.
When there are more examples (recipes and tiddlers), there may be.

Note that I have no enabled PATCH for tiddlers until /recipes. Unless
you know which tiddler you're working with you really shouldn't
be PATCHing it.

Thus far I'm still not feeling any "ZOMG this is awesome" about PATCH.